requirements.txt
moviepy==1.0.3
python-dotenv==1.0.1
boto3>=1.35.86
pydub==0.25.1
opencv-python==4.10.0.84
scikit-image==0.25.0
gradio==4.44.1
